Start the beginnings of czinspect, the intended successor to extractjxr.
czinspect is intended to be a more general and flexible Zeiss CZI handling
tool than extractjxr, which was originally written to provide a minimally
functional utility as soon as possible. As a result, extractjxr is not
exactly portable, requiring a 64-bit little endian machine to function
properly. czinspect is intended to support both big endian and 32-bit
machines and to ensure a reasonable degree of portability accross the
Unix-like landscape.

Immediate implementation notes:
 - The user interface is intended to be similar to Arch Linux's pacman,
   i.e. an operation is specified with a capital latin letter, and then
   suboptions specific to the operation are specified with a lower case
   latin letter.
 - The help text is contained in separate text files, and some linker
   tricks are used to convert these files into object files which can
   be linked into the final executable. Pointers to the start and end of
   these ELF sections can then be acquired, so the file contents can be
   accessed properly. This is probably going to come back and bite me at
   some point, but for now it's a neat hack which makes modifying help
   text (especially during development) nicer than messing around with
   fprintf() lines.